requirements.txt
aiogram
mongoengine
pyyaml